chore(audiobook): delete dead mini-player/pill/cover views superseded by the morphing player (#1233)
AudiobookMorphingPlayerView is the only mounted audiobook player surface since
the AppTabHostView keeper removal, so AudiobookMiniPlayerView,
AudiobookCollapsedPillView, and AudiobookFullPlayerCoverContainer render nowhere.
Delete them and their now-inert tests; relocate the still-used formatTime helper
(and its test) onto AudiobookMorphingPlayerView; drop stale doc comments that
referenced the removed types. Removed from both Palace and Palace-noDRM targets
via the xcodeproj gem.

**Scope:** dead-view deletion + formatTime relocation only. No behavior change to
the live morphing player.
**Not done:** nothing deferred — this is the complete cleanup the PR #1230
architect review flagged.
**Deferred:** none.
