--- Palace/AppInfrastructure/AudiobookFullPlayerCoverContainer.swift
@@ -1,241 +1,0 @@
-//
-//  AudiobookFullPlayerCoverContainer.swift
-//  Palace
-//
-//  Module D (swarm_0b7616e7) — root-level `fullScreenCover` content
-//  rendered by `AppTabHostView` when `presenter.isPlayerExpanded == true`.
-//  Hosts the existing `AudiobookPlayerView` and wraps it in a custom
-//  swipe-down gesture so the user can dismiss to the mini-player.
-//
-//  Swipe-down threshold: 100pt vertical translation with horizontal
-//  drift under 60pt (filters out diagonal scrolls so legitimate vertical
-//  swipes inside the player don't accidentally minimize). Reduce-motion
-//  callers skip the implicit animation; everyone else gets `.easeInOut`.
-//
-//  When `presenter.playbackModel == nil` the container renders
-//  `EmptyView()` so the cover never appears with an empty player —
-//  belt-and-suspenders against a race where `isPlayerExpanded == true`
-//  ran ahead of `adoptPlaybackModel(_:)` during a fast first-open.
-//
-//  Polish-phase additions (in-app-nav-polish-2026-06-01):
-//
-//    - Bug 1 fix: explicit top-leading chevron-down/Done button overlay
-//      that calls `presenter.minimize()`. iOS users reported the player
-//      had "no escape" because the existing 100pt swipe-down gesture
-//      was not discoverable. The chevron sits in the safe area's top
-//      inset, follows the `.regularMaterial` system convention for
-//      navigation chrome, and is announced as "Dismiss player" to
-//      VoiceOver.
-//    - VoiceOver focus is moved into the player on expansion via
-//      `UIAccessibility.post(.layoutChanged, argument: nil)` (announces
-//      a layout change so the focus engine re-acquires from the
-//      now-visible chrome — `@AccessibilityFocusState` bound through
-//      the toolkit's `AudiobookPlayerView` isn't reachable since that
-//      view is in a separate framework). The post is wrapped in a guard
-//      against VoiceOver-off so non-VO users aren't impacted.
-//
-//  Copyright (c) 2026 The Palace Project. All rights reserved.
-//
-
-import PalaceAudiobookToolkit
-import SwiftUI
-import UIKit
-
-@MainActor
-struct AudiobookFullPlayerCoverContainer: View {
-
-    @ObservedObject var presenter: AudiobookSessionPresenter
-
-    /// Swipe-down threshold in points. Exposed for tests so the boundary
-    /// case (just-under / just-over the threshold) can be asserted
-    /// against the same constant the production gesture uses.
-    static let minimizeSwipeDownThreshold: CGFloat = 100
-
-    /// Maximum horizontal drift before a downward swipe stops counting
-    /// as "vertical." Anything more than this is treated as a diagonal
-    /// scroll and ignored. Filters out content-scroll gestures inside
-    /// the player chrome.
-    static let minimizeSwipeMaxHorizontalDrift: CGFloat = 60
-
-    /// Drives the Table-of-Contents / Bookmarks `fullScreenCover`. The
-    /// toolkit's own TOC button is a `.navigationBar` toolbar item +
-    /// `NavigationLink` that only works inside a `NavigationStack` — which
-    /// this persistent overlay deliberately does NOT provide (a nested
-    /// `NavigationStack` inside the `TabView`'s nav stacks triggers SwiftUI's
-    /// "NavigationRequestObserver tried to update multiple times per frame"
-    /// fault and crashes on back-nav). Instead we present the toolkit's
-    /// `AudiobookNavigationView` (Chapters + Bookmarks) in an isolated
-    /// `fullScreenCover`, which has its own clean navigation context.
-    @State private var showTableOfContents = false
-
-    /// Live finger-tracking offset for the dismiss drag. Updated 1:1 (with
-    /// rubber-band resistance on upward drags) in the gesture's `.onChanged`,
-    /// and sprung back to 0 (or into the minimize) on `.onEnded`.
-    @State private var dragOffset: CGFloat = 0
-
-    @ViewBuilder
-    var body: some View {
-        if let model = presenter.playbackModel {
-            ZStack(alignment: .top) {
-                AudiobookPlayerView(model: model)
-                    .gesture(swipeDownToMinimize)
-                topControlsOverlay
-            }
-            // Track the finger during the dismiss drag (no implicit animation so
-            // it follows 1:1); the spring on release is applied in handleDragEnd.
-            .offset(y: dragOffset)
-            .fullScreenCover(isPresented: $showTableOfContents) {
-                NavigationStack {
-                    AudiobookNavigationView(model: model)
-                }
-            }
-            .onAppear {
-                postVoiceOverLayoutChangeIfNeeded()
-            }
-            .onChange(of: presenter.isPlayerExpanded) { _, expanded in
-                // Push focus into the cover on expand → true, and back
-                // out (system handles) on expand → false. The same
-                // `layoutChanged` post tells VoiceOver to re-acquire from
-                // the now-visible chrome.
-                if expanded { postVoiceOverLayoutChangeIfNeeded() }
-            }
-        } else {
-            EmptyView()
-        }
-    }
-
-    /// Top chrome for the full player: the non-destructive chevron-down
-    /// Done button (leading, minimizes to the mini-player) and the
-    /// Table-of-Contents button (trailing, presents Chapters/Bookmarks).
-    /// Mirrors the legacy player's leading-back / trailing-TOC layout
-    /// without the toolkit's `.navigationBar` toolbar (which needs a
-    /// `NavigationStack` the overlay can't safely host).
-    private var topControlsOverlay: some View {
-        HStack {
-            doneButton
-            Spacer()
-            tableOfContentsButton
-        }
-        .padding(.horizontal, 12)
-        .padding(.top, 12)
-        .frame(maxWidth: .infinity, alignment: .top)
-    }
-
-    /// Leading chevron-down Done button (Bug 1 fix —
-    /// in-app-nav-polish-2026-06-01). iOS HIG convention for dismissing a
-    /// sheet/cover is top-leading "Done"; we use chevron-down because the
-    /// matching gesture is also vertical (swipe down), giving a consistent
-    /// "going down = dismiss" mental model. Positioning is handled by
-    /// `topControlsOverlay`.
-    private var doneButton: some View {
-        Button(action: minimizeWithMotionPreference) {
-            Image(systemName: "chevron.down")
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 18, height: 18)
-                .padding(12)
-                .background(.regularMaterial, in: Circle())
-        }
-        .buttonStyle(.plain)
-        .tint(.primary)
-        .frame(width: 44, height: 44)
-        .accessibilityLabel(Strings.Generic.dismissPlayer)
-        .accessibilityAddTraits(.isButton)
-    }
-
-    /// Trailing Table-of-Contents button. Presents the toolkit's
-    /// `AudiobookNavigationView` (Chapters + Bookmarks) via the
-    /// `showTableOfContents` `fullScreenCover` — the overlay's
-    /// nested-`NavigationStack`-free replacement for the toolkit's own
-    /// `.navigationBar` TOC button.
-    private var tableOfContentsButton: some View {
-        Button { showTableOfContents = true } label: {
-            Image(systemName: "list.bullet")
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 20, height: 20)
-                .padding(12)
-                .background(.regularMaterial, in: Circle())
-        }
-        .buttonStyle(.plain)
-        .tint(.primary)
-        .frame(width: 44, height: 44)
-        .accessibilityLabel(Strings.Generic.tableOfContents)
-        .accessibilityAddTraits(.isButton)
-    }
-
-    private var swipeDownToMinimize: some Gesture {
-        DragGesture(minimumDistance: 10, coordinateSpace: .local)
-            .onChanged { value in
-                // Follow the finger live (rubber-banded upward) so the drag
-                // feels direct instead of only reacting on release.
-                dragOffset = Self.rubberBandedDragOffset(translationHeight: value.translation.height)
-            }
-            .onEnded { value in
-                handleDragEnd(translation: value.translation)
-            }
-    }
-
-    /// Test-visible drag-end handler. The test seam takes a raw
-    /// `CGSize` so unit tests can simulate any drag without spinning
-    /// up a real DragGesture. Production calls this from the
-    /// `.onEnded` closure above. Springs the tracked offset back to rest
-    /// (or into the minimize) — the same threshold contract as before, now
-    /// resolved through the pure `shouldMinimize(translation:)` seam.
-    func handleDragEnd(translation: CGSize) {
-        let minimize = Self.shouldMinimize(translation: translation)
-        if UIAccessibility.isReduceMotionEnabled {
-            dragOffset = 0
-            if minimize { presenter.minimize() }
-        } else {
-            withAnimation(PalaceMotion.emphasized) {
-                dragOffset = 0
-                if minimize { presenter.minimize() }
-            }
-        }
-    }
-
-    /// Rubber-banded finger-tracking offset for the dismiss drag. Downward
-    /// translation is followed 1:1; upward translation is heavily resisted
-    /// (15%) so the player can't be dragged up past its resting position.
-    static func rubberBandedDragOffset(translationHeight: CGFloat) -> CGFloat {
-        translationHeight >= 0 ? translationHeight : translationHeight * 0.15
-    }
-
-    /// Pure dismiss decision: a downward drag past the threshold with limited
-    /// horizontal drift. Identical contract to the guard the `.onEnded` handler
-    /// used before the finger-tracking upgrade, extracted so the boundary stays
-    /// unit-testable.
-    static func shouldMinimize(translation: CGSize) -> Bool {
-        translation.height > minimizeSwipeDownThreshold
-            && abs(translation.width) < minimizeSwipeMaxHorizontalDrift
-    }
-
-    /// Drives `presenter.minimize()` honoring the user's reduce-motion
-    /// preference. Wrapping in `withAnimation` adds an implicit
-    /// easeInOut curve, which is exactly what reduce-motion users have
-    /// asked the system NOT to do.
-    private func minimizeWithMotionPreference() {
-        if UIAccessibility.isReduceMotionEnabled {
-            presenter.minimize()
-        } else {
-            withAnimation(.easeInOut) { presenter.minimize() }
-        }
-    }
-
-    /// Posts a `.layoutChanged` UIAccessibility notification so the
-    /// VoiceOver focus engine re-acquires from the newly-presented
-    /// player chrome. Guarded on `UIAccessibility.isVoiceOverRunning`
-    /// so non-VO users aren't impacted by an unnecessary post.
-    ///
-    /// `argument: nil` lets the system pick a sensible focus target
-    /// (usually the first focusable element in the now-visible
-    /// hierarchy). `@AccessibilityFocusState` would be the SwiftUI-
-    /// native path, but the focusable target lives inside
-    /// `AudiobookPlayerView` (toolkit) which we don't own — posting
-    /// `layoutChanged` is the cross-framework-safe alternative.
-    private func postVoiceOverLayoutChangeIfNeeded() {
-        guard UIAccessibility.isVoiceOverRunning else { return }
-        UIAccessibility.post(notification: .layoutChanged, argument: nil)
-    }
-}
