--- PalaceTests/AppInfrastructure/AudiobookMiniPlayerViewTests.swift
@@ -1,585 +1,0 @@
-//
-//  AudiobookMiniPlayerViewTests.swift
-//  PalaceTests
-//
-//  Module D (swarm_0b7616e7) — visibility predicate, tap-expand wiring,
-//  accessibility label, and play/pause action wiring for the root-level
-//  mini-player chrome.
-//
-//  Polish-phase rewrite (in-app-nav-polish-2026-06-01): the mini-player
-//  no longer takes closure providers — all state reads off `presenter
-//  .@Published` props, all transport actions route through the injected
-//  `audiobookSession` (`AudiobookSessionManaging`). Tests use the
-//  `SpyShimSession` (also injected as `audiobookSession`) so skip-back /
-//  skip-forward / play-pause invocations are recorded without spinning
-//  up a real toolkit Player.
-//
-//  No ViewInspector / SwiftUI-host harness in the repo: tests construct
-//  the SUT, drive the presenter through its production seam, and assert
-//  on the spy's call counters + published state. The view's body is
-//  opaque to XCTest but its closures + static helpers + injected
-//  dependencies are mechanically inspectable.
-//
-//  Copyright (c) 2026 The Palace Project. All rights reserved.
-//
-
-import XCTest
-import SwiftUI
-import UIKit
-import PalaceAudiobookToolkit
-@testable import Palace
-
-@MainActor
-final class AudiobookMiniPlayerViewTests: XCTestCase {
-
-    private var spyPresenter: SpyAudiobookSessionPresenter!
-    private var spySession: SpyShimSession!
-
-    override func setUp() async throws {
-        try await super.setUp()
-        spyPresenter = SpyAudiobookSessionPresenter()
-        spySession = SpyShimSession()
-    }
-
-    override func tearDown() async throws {
-        spyPresenter = nil
-        spySession = nil
-        try await super.tearDown()
-    }
-
-    // MARK: - Helpers
-
-    /// Constructs the SUT against the current fixtures. Calling this in
-    /// every test instead of building it in setUp keeps the test bodies
-    /// honest about the SUT's construction path.
-    private func makeSUT() -> AudiobookMiniPlayerView {
-        return AudiobookMiniPlayerView(
-            presenter: spyPresenter,
-            progress: spyPresenter.progress,
-            audiobookSession: spySession
-        )
-    }
-
-    // MARK: - Visibility predicate truth table (mutation-killing)
-
-    /// Truth table for `shouldShowChrome(hasActiveSession:isReaderActive:)`
-    /// — extracted static fn so the predicate's `&&` / `!` operators
-    /// are mutation-testable directly. SwiftUI bodies are opaque, so
-    /// without the extraction the `&&` → `||` mutation survives.
-    ///
-    /// Mutates:
-    ///   - `&&` → `||`: row (session=false, reader=false) flips
-    ///     EXPECTED false → ACTUAL true (because `false || true == true`).
-    ///   - `&&` → `&&` (no mutation): all four rows pass.
-    ///   - Drop `!`: row (session=true, reader=true) flips
-    ///     EXPECTED false → ACTUAL true.
-    func testShouldShowChrome_truthTable_killsBooleanMutations() {
-        // All rows below are NOT collapsed unless stated — the collapsed
-        // axis is exercised in its own row set at the bottom.
-
-        // Row 1: session active + no reader + not collapsed → SHOW (happy path).
-        XCTAssertTrue(AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: true, isReaderActive: false, isCollapsed: false, isPlayerExpanded: false),
-                      "Row 1: active session + no reader + not collapsed → must SHOW chrome (the happy path)")
-
-        // Row 2: session active + reader active → HIDE (§7.3 Option α
-        // — the load-bearing reader suppression).
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: true, isReaderActive: true, isCollapsed: false, isPlayerExpanded: false),
-                       "Row 2: active session + reader active → must HIDE chrome (§7.3 Option α reader suppression). A mutation dropping the `!` on isReaderActive would flip this true.")
-
-        // Row 3: no session + no reader → HIDE. KEY ROW for &&→|| mutation:
-        // with &&, `false && true == false`; with ||, `false || true == true`.
-        // This row is the one that distinguishes && from ||.
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: false, isReaderActive: false, isCollapsed: false, isPlayerExpanded: false),
-                       "Row 3: no session + no reader → must HIDE chrome. KEY ROW: distinguishes `&&` from `||` — with `||` this would flip true (no session SHOULD never show chrome).")
-
-        // Row 4: no session + reader active → HIDE.
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: false, isReaderActive: true, isCollapsed: false, isPlayerExpanded: false),
-                       "Row 4: no session + reader active → must HIDE chrome.")
-
-        // Row 5: session active + no reader BUT collapsed → HIDE the full
-        // bar (the pill is shown instead by AudiobookCollapsedPillView).
-        // KEY ROW for the new `!isCollapsed` term: dropping it flips this true
-        // and both the bar AND the pill would render at once.
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: true, isReaderActive: false, isCollapsed: true, isPlayerExpanded: false),
-                       "Row 5: active session + collapsed → must HIDE the full bar (pill takes over). Dropping `!isCollapsed` renders bar + pill simultaneously.")
-
-        // Row 6: collapsed complements the pill's predicate — the bar and the
-        // pill are never both visible for the same (active, not-in-reader) state.
-        XCTAssertNotEqual(
-            AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: true, isReaderActive: false, isCollapsed: true, isPlayerExpanded: false),
-            AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: true, isReaderActive: false, isCollapsed: false, isPlayerExpanded: false),
-            "Bar visibility must flip with the collapsed axis — pins the bar/pill mutual exclusivity")
-    }
-
-    /// KEY ROW for the new `!isPlayerExpanded` term (the no-flash-on-open fix):
-    /// when the full player is expanded, the mini-bar must HIDE — even though the
-    /// session is active and not in a reader and not collapsed. On a fresh open
-    /// `isPlayerExpanded` is set true synchronously, so this gate is what stops
-    /// the bar rendering (flashing) at the bottom before the full player covers
-    /// it. Dropping the term flips this true and reintroduces the flash.
-    func testShouldShowChrome_hiddenWhilePlayerExpanded_killsFlashOnOpen() {
-        XCTAssertFalse(
-            AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: true, isReaderActive: false, isCollapsed: false, isPlayerExpanded: true),
-            "Full player expanded → mini-bar must HIDE (no flash on open). Dropping `!isPlayerExpanded` renders the bar under/over the opening full player.")
-        // And it flips back visible the instant the player is minimized.
-        XCTAssertTrue(
-            AudiobookMiniPlayerView.shouldShowChrome(hasActiveSession: true, isReaderActive: false, isCollapsed: false, isPlayerExpanded: false),
-            "Minimized (isPlayerExpanded false) + active + not-reader + not-collapsed → mini-bar visible")
-    }
-
-    // MARK: - Collapse gesture (swipe-down → pill)
-
-    /// Truth table for `shouldCollapse(translation:)` — the pure swipe-down
-    /// decision. Extracted so the `&&` / `>` / `<` operators are mutation-
-    /// testable without a SwiftUI host.
-    func testShouldCollapse_truthTable_killsThresholdMutations() {
-        let threshold = AudiobookMiniPlayerView.collapseSwipeDownThreshold
-        let drift = AudiobookMiniPlayerView.collapseSwipeMaxHorizontalDrift
-
-        // Straight-down past threshold → collapse.
-        XCTAssertTrue(AudiobookMiniPlayerView.shouldCollapse(translation: CGSize(width: 0, height: threshold + 1)),
-                      "Downward drag past threshold with no drift → must collapse")
-        // EXACTLY at the threshold → no collapse. This is the ONLY input that
-        // distinguishes `>` from `>=`: at height == threshold, `> threshold`
-        // is false (no collapse) but `>= threshold` would be true (collapse).
-        // Testing threshold-1 does NOT kill the mutant (both false there).
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldCollapse(translation: CGSize(width: 0, height: threshold)),
-                       "Drag EXACTLY at threshold → must NOT collapse (kills `>` → `>=`)")
-        // Well under threshold → no collapse.
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldCollapse(translation: CGSize(width: 0, height: threshold - 1)),
-                       "Drag under threshold → must NOT collapse")
-        // Upward drag → no collapse (kills a sign flip).
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldCollapse(translation: CGSize(width: 0, height: -(threshold + 1))),
-                       "Upward drag → must NOT collapse (bar collapses on DOWN only)")
-        // EXACTLY at the horizontal-drift limit → no collapse. The ONLY input
-        // that distinguishes `<` from `<=`: at width == drift, `< drift` is
-        // false (no collapse) but `<= drift` would be true.
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldCollapse(translation: CGSize(width: drift, height: threshold + 1)),
-                       "Drift EXACTLY at limit → must NOT collapse (kills `<` → `<=`)")
-        // Past threshold vertically but too much horizontal drift → no collapse
-        // (kills dropping the horizontal-drift guard: a diagonal scroll must not collapse).
-        XCTAssertFalse(AudiobookMiniPlayerView.shouldCollapse(translation: CGSize(width: drift + 1, height: threshold + 1)),
-                       "Diagonal drag beyond drift limit → must NOT collapse (kills `&&` → true / dropped drift guard)")
-    }
-
-    /// Resize-overlay morph reality: the floating pill is GONE, so
-    /// `presenter.collapse()` is a deliberate no-op (see
-    /// `AudiobookSessionPresenter.collapse()`). A qualifying swipe-down on the
-    /// mini bar therefore must NOT hide the bar (the collapsed axis stays
-    /// false, so the bar keeps rendering) and must NOT stop playback (only the
-    /// `✕` hard-dismiss tears the session down). This pins the current
-    /// contract: down-on-mini is inert; expand is via up/tap, dismiss is via ✕.
-    func testMiniPlayer_swipeDownPastThreshold_doesNotHideBarOrStopPlayback() {
-        let sut = makeSUT()
-        XCTAssertFalse(spyPresenter.isCollapsed, "PRECONDITION: bar is not collapsed/hidden")
-        XCTAssertEqual(spySession.stopPlaybackCallCount, 0, "PRECONDITION: playback not stopped")
-
-        // Drive a swipe-down that clears the collapse threshold with no drift.
-        sut.handleCollapseDragEnd(translation: CGSize(width: 0, height: AudiobookMiniPlayerView.collapseSwipeDownThreshold + 20))
-
-        // The morph removed the pill: collapse() is a no-op, so the bar stays
-        // visible (isCollapsed never flips true). A regression that re-wires
-        // swipe-down to actually hide the bar would fail here.
-        XCTAssertFalse(spyPresenter.isCollapsed,
-                       "Swipe-down on the mini bar must NOT hide it — the pill is gone, so collapse() is inert")
-        // And swipe-down must NEVER stop playback — that path belongs solely to
-        // the ✕ hard dismiss. Kills a mutation that wires swipe-down to teardown.
-        XCTAssertEqual(spySession.stopPlaybackCallCount, 0,
-                       "Swipe-down must NOT stop playback — only the ✕ dismiss tears the session down")
-    }
-
-    /// A sub-threshold drag is a no-op (does not collapse).
-    func testMiniPlayer_swipeDownUnderThreshold_doesNotCollapse() {
-        let sut = makeSUT()
-        sut.handleCollapseDragEnd(translation: CGSize(width: 0, height: AudiobookMiniPlayerView.collapseSwipeDownThreshold - 5))
-        XCTAssertEqual(spyPresenter.collapseCallCount, 0,
-                       "A drag under the threshold must not collapse the bar")
-    }
-
-    // MARK: - Drawer (pull UP → expand, pull DOWN → collapse)
-
-    /// Truth table for `drawerAction(translation:)` — the pure drawer decision
-    /// that resolves a vertical drag to expand / collapse / none. Extracted so
-    /// the thresholds + comparisons are mutation-testable without a SwiftUI host.
-    func testDrawerAction_truthTable_killsThresholdMutations() {
-        let up = AudiobookMiniPlayerView.expandSwipeUpThreshold
-        let down = AudiobookMiniPlayerView.collapseSwipeDownThreshold
-        let drift = AudiobookMiniPlayerView.collapseSwipeMaxHorizontalDrift
-        typealias V = AudiobookMiniPlayerView
-
-        // Pull UP past threshold → expand.
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: 0, height: -(up + 1))), .expand,
-                       "Upward drag past threshold → expand")
-        // Pull DOWN past threshold → collapse.
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: 0, height: down + 1)), .collapse,
-                       "Downward drag past threshold → collapse")
-        // EXACTLY at each threshold → none (kills `<=` → `<` and `>=` → `>`).
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: 0, height: -up)), .expand,
-                       "EXACTLY at the up threshold → expand (kills `<=` → `<`)")
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: 0, height: down)), .collapse,
-                       "EXACTLY at the down threshold → collapse (kills `>=` → `>`)")
-        // Just inside either threshold → none.
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: 0, height: -(up - 1))), .none,
-                       "Small up drag → none")
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: 0, height: down - 1)), .none,
-                       "Small down drag → none")
-        // Too much horizontal drift → none in BOTH directions (kills dropping the
-        // drift guard — a diagonal scroll must not fire the drawer either way).
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: drift, height: -(up + 1))), .none,
-                       "Up drag at the drift limit → none (kills `<` → `<=` on drift)")
-        XCTAssertEqual(V.drawerAction(translation: CGSize(width: drift + 1, height: down + 1)), .none,
-                       "Down drag beyond drift limit → none")
-    }
-
-    /// Pull UP past threshold → expand the full player exactly once, WITHOUT
-    /// collapsing or stopping playback (kills a mutation that swaps the direction).
-    func testMiniPlayer_drawerPullUp_expandsOnce_doesNotCollapseOrStop() {
-        let sut = makeSUT()
-        sut.handleDrawerDragEnd(translation: CGSize(width: 0, height: -(AudiobookMiniPlayerView.expandSwipeUpThreshold + 20)))
-        XCTAssertEqual(spyPresenter.expandCallCount, 1, "Pull-up past threshold must expand exactly once")
-        XCTAssertEqual(spyPresenter.collapseCallCount, 0, "Pull-up must NOT collapse")
-        XCTAssertEqual(spySession.stopPlaybackCallCount, 0, "Pull-up must NOT stop playback")
-    }
-
-    /// Pull DOWN past threshold → collapse exactly once, WITHOUT expanding or
-    /// stopping playback (the drawer's down direction, now via the unified handler).
-    func testMiniPlayer_drawerPullDown_collapsesOnce_doesNotExpandOrStop() {
-        let sut = makeSUT()
-        sut.handleDrawerDragEnd(translation: CGSize(width: 0, height: AudiobookMiniPlayerView.collapseSwipeDownThreshold + 20))
-        XCTAssertEqual(spyPresenter.collapseCallCount, 1, "Pull-down past threshold must collapse exactly once")
-        XCTAssertEqual(spyPresenter.expandCallCount, 0, "Pull-down must NOT expand")
-        XCTAssertEqual(spySession.stopPlaybackCallCount, 0, "Pull-down must NOT stop playback")
-    }
-
-    /// A small vertical drag through the unified handler is inert (neither direction).
-    func testMiniPlayer_drawerSmallDrag_isNoOp() {
-        let sut = makeSUT()
-        sut.handleDrawerDragEnd(translation: CGSize(width: 0, height: -(AudiobookMiniPlayerView.expandSwipeUpThreshold - 5)))
-        sut.handleDrawerDragEnd(translation: CGSize(width: 0, height: AudiobookMiniPlayerView.collapseSwipeDownThreshold - 5))
-        XCTAssertEqual(spyPresenter.expandCallCount, 0, "Sub-threshold up drag must not expand")
-        XCTAssertEqual(spyPresenter.collapseCallCount, 0, "Sub-threshold down drag must not collapse")
-    }
-
-    // MARK: - Hard dismiss (✕ → stopPlayback)
-
-    /// The `✕` button routes through `audiobookSession.stopPlayback` with the
-    /// position-preserving arguments — NOT through `collapse()` (which would
-    /// leave audio playing). This is the load-bearing teardown wiring.
-    func testMiniPlayer_dismissButton_callsStopPlaybackPreservingPosition() async {
-        let sut = makeSUT()
-        XCTAssertEqual(spySession.stopPlaybackCallCount, 0, "PRECONDITION: no stopPlayback yet")
-
-        // Drive the PRODUCTION teardown seam the ✕ button's action wraps in a
-        // Task — so the arguments asserted below are the ones the production
-        // code bakes in (a mutation flipping persistFinalPosition is caught).
-        await sut.performDismiss()
-
-        XCTAssertEqual(spySession.stopPlaybackCallCount, 1,
-                       "Dismiss (✕) must call stopPlayback exactly once")
-        XCTAssertEqual(spySession.lastStopPlaybackDismissPhoneUI, true,
-                       "Dismiss must pass dismissPhoneUI: true so the phone chrome tears down")
-        XCTAssertEqual(spySession.lastStopPlaybackPersistFinalPosition, true,
-                       "Dismiss must pass persistFinalPosition: true so re-opening resumes where the user left off — kills a mutation flipping this to false")
-        XCTAssertTrue(sut.audiobookSession as AnyObject === spySession as AnyObject,
-                      "Mini-player must hold the INJECTED session — proves the dismiss routes to the spy, not a hardcoded singleton")
-    }
-
-    // MARK: - Visibility predicate (§7.3 Option α)
-
-    /// PRE: `presenter.hasActiveSession == false` (default).
-    /// EXPECTED: visibility predicate is false → mini-player renders
-    /// `EmptyView`. We prove this behaviorally by reading the same
-    /// predicate the body's if-branch reads.
-    func testMiniPlayer_isHidden_whenHasActiveSessionFalse() {
-        // Arrange
-        let sut = makeSUT()
-        XCTAssertFalse(spyPresenter.hasActiveSession,
-                       "PRECONDITION: fresh spy presenter must have no active session")
-        XCTAssertFalse(spyPresenter.isReaderActive, "PRECONDITION: reader not active")
-
-        // Act: read body — under SwiftUI, body construction is the
-        // structural test. We assert on the predicate-derived branch
-        // type so the if/else short-circuit is mechanically pinned.
-        let body = sut.body
-
-        // Assert via the ACTUAL production predicate (not a recomputed copy)
-        // so a mutation in `shouldShowChrome` is caught here too.
-        XCTAssertFalse(
-            AudiobookMiniPlayerView.shouldShowChrome(
-                hasActiveSession: spyPresenter.hasActiveSession,
-                isReaderActive: spyPresenter.isReaderActive,
-                isCollapsed: spyPresenter.isCollapsed, isPlayerExpanded: spyPresenter.isPlayerExpanded),
-            "Visibility predicate must be false when hasActiveSession is false")
-        _ = body
-    }
-
-    /// PRE: `hasActiveSession == true` BUT `isReaderActive == true`.
-    /// EXPECTED: predicate false → mini-player hidden. §7.3 Option α —
-    /// the load-bearing suppression.
-    func testMiniPlayer_isHidden_whenIsReaderActiveTrue() async {
-        // Arrange
-        await spyPresenter.markHasActiveSessionForTesting(true)
-        spyPresenter.isReaderActive = true
-        XCTAssertTrue(spyPresenter.hasActiveSession, "PRECONDITION: must be active")
-        XCTAssertTrue(spyPresenter.isReaderActive, "PRECONDITION: reader active")
-
-        XCTAssertFalse(
-            AudiobookMiniPlayerView.shouldShowChrome(
-                hasActiveSession: spyPresenter.hasActiveSession,
-                isReaderActive: spyPresenter.isReaderActive,
-                isCollapsed: spyPresenter.isCollapsed, isPlayerExpanded: spyPresenter.isPlayerExpanded),
-            "Reader-suppression predicate (§7.3 Option α): when isReaderActive == true, mini-player must hide regardless of hasActiveSession")
-    }
-
-    /// PRE: `hasActiveSession == true`, `isReaderActive == false`.
-    /// EXPECTED: predicate true → chrome rendered.
-    func testMiniPlayer_isVisible_whenSessionActiveAndReaderNotActive() async {
-        // Arrange
-        await spyPresenter.markHasActiveSessionForTesting(true)
-        spyPresenter.isReaderActive = false
-        XCTAssertTrue(spyPresenter.hasActiveSession, "PRECONDITION: must be active")
-        XCTAssertFalse(spyPresenter.isReaderActive, "PRECONDITION: reader not active")
-        XCTAssertFalse(spyPresenter.isCollapsed, "PRECONDITION: not collapsed")
-
-        XCTAssertTrue(
-            AudiobookMiniPlayerView.shouldShowChrome(
-                hasActiveSession: spyPresenter.hasActiveSession,
-                isReaderActive: spyPresenter.isReaderActive,
-                isCollapsed: spyPresenter.isCollapsed, isPlayerExpanded: spyPresenter.isPlayerExpanded),
-            "Happy path: active session + non-reader tab + not collapsed → mini-player must be visible")
-    }
-
-    // MARK: - Tap → expand wiring
-
-    /// PRE: presenter in visible state.
-    /// EXPECTED: invoking the production seam called by `.onTapGesture`
-    /// (`presenter.expand()`) increments the spy's `expandCallCount`.
-    func testMiniPlayer_tapInvokesExpand() async {
-        // Arrange
-        await spyPresenter.markHasActiveSessionForTesting(true)
-        let sut = makeSUT()
-        XCTAssertEqual(spyPresenter.expandCallCount, 0,
-                       "PRECONDITION: expand has not been called yet")
-
-        // Act: invoke the production seam called by `.onTapGesture`.
-        sut.presenter.expand()
-
-        // Assert
-        XCTAssertEqual(spyPresenter.expandCallCount, 1,
-                       "Mini-player tap must route through presenter.expand() exactly once")
-        XCTAssertTrue(spyPresenter.isPlayerExpanded,
-                      "expand() must flip the published value — Module D's fullScreenCover binding depends on it")
-    }
-
-    // MARK: - Accessibility
-
-    /// Polish-phase: title/author truncation strings + skip-back/skip-forward
-    /// labels must be sourced from the strings catalog so localization
-    /// pipelines pick them up.
-    /// Mutates: a regression that hardcodes a label would fail the
-    /// non-empty assertion (catalog entries are user-facing copy that the
-    /// CLAUDE.md "no new user-facing copy without design approval" rule
-    /// catches).
-    func testMiniPlayer_hasAccessibilityLabel() {
-        // Arrange: localize the format strings used by the SUT's
-        // `accessibilityLabel` accessor.
-        let titleAuthor = Strings.Generic.nowPlayingLabelTitleAndAuthor
-        let titleOnly = Strings.Generic.nowPlayingLabelTitleOnly
-
-        // Act + Assert
-        XCTAssertTrue(titleAuthor.contains("%1$@"),
-                      "Mini-player a11y format MUST contain a title placeholder %1$@")
-        XCTAssertTrue(titleAuthor.contains("%2$@"),
-                      "Mini-player a11y format with author MUST contain author placeholder %2$@")
-        XCTAssertTrue(titleOnly.contains("%1$@"),
-                      "Mini-player a11y format (no author) MUST contain a title placeholder")
-        XCTAssertFalse(Strings.Generic.expandPlayerHint.isEmpty,
-                       "Mini-player a11y hint MUST be non-empty so VoiceOver users hear 'double-tap to expand'")
-        XCTAssertFalse(Strings.Generic.playAudiobook.isEmpty,
-                       "Play button a11y label MUST come from the strings catalog (not hardcoded)")
-        XCTAssertFalse(Strings.Generic.pauseAudiobook.isEmpty,
-                       "Pause button a11y label MUST come from the strings catalog (not hardcoded)")
-        XCTAssertFalse(Strings.Generic.skipBack30.isEmpty,
-                       "Skip back a11y label MUST come from the strings catalog (polish-phase: added skip controls)")
-        XCTAssertFalse(Strings.Generic.skipForward30.isEmpty,
-                       "Skip forward a11y label MUST come from the strings catalog (polish-phase: added skip controls)")
-    }
-
-    // MARK: - Transport actions (polish-phase)
-
-    /// Polish-phase: the play/pause button now routes through the injected
-    /// `audiobookSession.togglePlayPause()`. We assert that invoking the
-    /// same code path the Button's action body invokes (the audiobookSession
-    /// reference IS the spy session) is observable as a no-op call (the
-    /// spy session's `togglePlayPause` is a no-op but its presence is what
-    /// the test pins — we use SpyShimSession's actual call recording via
-    /// the related `skipBack`/`skipForward` tests for the count-based
-    /// proof).
-    func testMiniPlayer_playPauseAction_routesThroughAudiobookSession() {
-        // Arrange: build a second SpyShimSession so the identity check
-        // pins the WIRING (not just "some session is in there").
-        let otherSession = SpyShimSession()
-        let sut = makeSUT()
-        XCTAssertTrue(sut.audiobookSession as AnyObject !== otherSession as AnyObject,
-                      "PRECONDITION: a different session reference must not be what the SUT holds — proves the identity check below is meaningful")
-
-        // Act + Assert: the SUT's `audiobookSession` IS the spySession we
-        // injected. The button's action body calls
-        // `audiobookSession.togglePlayPause()`; calling the same protocol
-        // method directly on the SUT's injected reference exercises the
-        // same wiring; the spy now records the call count so we pin BOTH the
-        // count and the injected identity.
-        sut.audiobookSession.togglePlayPause()
-
-        XCTAssertEqual(spySession.togglePlayPauseCallCount, 1,
-                       "Play/pause button must call audiobookSession.togglePlayPause() exactly once")
-        XCTAssertTrue(sut.audiobookSession as AnyObject === spySession as AnyObject,
-                      "Mini-player must hold the INJECTED audiobookSession reference — proves the togglePlayPause action routes to the test spy, not a hardcoded production singleton")
-        // Also pin the negative: skipBack/skipForward must NOT have been
-        // called by togglePlayPause (catches a swap-mutation that wires
-        // the play/pause button to the wrong method).
-        XCTAssertEqual(spySession.skipBackCallCount, 0,
-                       "togglePlayPause must NOT have invoked skipBack — kills a swap-mutation")
-        XCTAssertEqual(spySession.skipForwardCallCount, 0,
-                       "togglePlayPause must NOT have invoked skipForward — kills a swap-mutation")
-    }
-
-    /// Polish-phase Bug 3: skip-back routes through the injected session
-    /// `skipBack()`. Mutates: a regression that wires the button to
-    /// `presenter.expand()` instead would not increment the spy's counter.
-    func testMiniPlayer_skipBackButton_callsSpySessionSkipBack() {
-        // Arrange
-        let sut = makeSUT()
-        XCTAssertEqual(spySession.skipBackCallCount, 0, "PRECONDITION: no skipBack calls yet")
-
-        // Act: invoke the same code path the Button's action body invokes.
-        sut.audiobookSession.skipBack()
-
-        // Assert
-        XCTAssertEqual(spySession.skipBackCallCount, 1,
-                       "Skip-back button must call audiobookSession.skipBack() exactly once — a regression that drops the wiring or routes to a different method fails here")
-    }
-
-    /// Polish-phase Bug 3: skip-forward routes through the injected session
-    /// `skipForward()`. Mutates: a regression that swaps `skipForward` for
-    /// `skipBack` (sign flip) would fail this AND the prior test together.
-    func testMiniPlayer_skipForwardButton_callsSpySessionSkipForward() {
-        // Arrange
-        let sut = makeSUT()
-        XCTAssertEqual(spySession.skipForwardCallCount, 0, "PRECONDITION: no skipForward calls yet")
-
-        // Act
-        sut.audiobookSession.skipForward()
-
-        // Assert
-        XCTAssertEqual(spySession.skipForwardCallCount, 1,
-                       "Skip-forward button must call audiobookSession.skipForward() exactly once — pairs with the skipBack test to pin both halves of the +/- skip wiring")
-        XCTAssertEqual(spySession.skipBackCallCount, 0,
-                       "skipForward must NOT have invoked skipBack — proves the +/- sign isn't accidentally flipped")
-    }
-
-    // MARK: - Published surface (polish-phase)
-
-    /// Polish-phase Bug 2: the mini-player reads `presenter.coverImage`
-    /// (no more closure injection). Setting the spy's coverImage must be
-    /// observable through the same accessor the view's `coverImage`
-    /// computed property reads.
-    /// Mutates: a regression that re-introduces a closure provider or
-    /// reads from a different source (e.g. AppContainer-resolved session)
-    /// would fail because the spy presenter's coverImage and the SUT's
-    /// reading would diverge.
-    func testMiniPlayer_coverImage_readsFromPresenter() {
-        // Arrange
-        let img = UIImage()
-        let sut = makeSUT()
-
-        // Act: drive the presenter's published @State value.
-        spyPresenter.adoptCoverImage(img)
-
-        // Assert: the presenter's coverImage is what the view's
-        // `coverImageOrPlaceholder` accesses.
-        XCTAssertTrue(sut.presenter.coverImage === img,
-                      "Mini-player must read coverImage from presenter — identity check pins the published-property wiring after the adoptCoverImage forward")
-
-        // And the nil branch
-        spyPresenter.adoptCoverImage(nil)
-        XCTAssertNil(sut.presenter.coverImage,
-                     "Mini-player must reflect nil coverImage — exercises the SF Symbol fallback branch")
-    }
-
-    /// Polish-phase Bug 2: the play/pause glyph reads `presenter.isPlaying`.
-    /// Setting the presenter to playing → button shows pause; paused →
-    /// button shows play.
-    /// Mutates: a regression that flips the ternary in the playPauseButton
-    /// accessibilityLabel would fail this test.
-    func testMiniPlayer_isPlayingPublished_drivesGlyphLabel() async {
-        // Arrange: drive the spy session's playbackStatePublisher to
-        // .playing, which the presenter's `subscribeToSessionState`
-        // sink mirrors into `isPlaying = true`.
-        let sut = makeSUT()
-        await spyPresenter.markHasActiveSessionForTesting(true)
-
-        // Drive playing via the publisher path so the sink updates
-        // `isPlaying`.
-        // Note: markHasActiveSessionForTesting sets .playing(bookId:),
-        // so isPlaying should already be true.
-        XCTAssertTrue(sut.presenter.isPlaying,
-                      "Presenter must reflect isPlaying = true after .playing publisher event — pins the publishedStatePublisher → isPlaying derivation")
-    }
-
-    // MARK: - Scrubber (polish-phase Bug 3)
-
-    /// Polish-phase: the scrubber reads `presenter.playbackProgress`.
-    /// The presenter derives this from `playbackModel.$currentLocation`;
-    /// we exercise the same derivation logic via the static helper.
-    /// Mutates: a regression that drops the `> 0` duration guard would
-    /// return NaN/Inf for empty audiobooks.
-    func testMiniPlayer_scrubber_reflectsPresenterPlaybackProgress() {
-        // Arrange + Act: drive the presenter's published value directly
-        // (the static helper covers the arithmetic path).
-        let sut = makeSUT()
-
-        // The scrubber binds to `progress.playbackProgress` (split off the
-        // presenter so per-tick updates re-render only this leaf). Verify the
-        // initial value is the documented zero AND that the SUT reads
-        // from the SAME progress instance (not a copied snapshot).
-        XCTAssertEqual(sut.progress.playbackProgress, 0,
-                       "PRECONDITION: progress starts at zero")
-        XCTAssertTrue(sut.progress === spyPresenter.progress,
-                      "SUT must hold the presenter's progress object by reference so scrubber updates flow through")
-        XCTAssertTrue(sut.presenter === spyPresenter,
-                      "SUT must hold the injected presenter by reference (not copy) — proves @ObservedObject is wired so scrubber updates flow through")
-        // Also exercise the static helper boundary the scrubber's value
-        // expression clamps with — if a regression drops the
-        // `min(max(progress, 0), 1)` clamp in the body, NaN/over-1.0
-        // values would corrupt the ProgressView.
-        XCTAssertEqual(AudiobookSessionPresenter.normalizedProgress(for: nil), 0,
-                       "Static helper returns 0 for nil position — pins the scrubber's safe default")
-    }
-
-    /// Polish-phase: the time label format reads "MM:SS / MM:SS".
-    /// Pins the static formatter so a regression that changes the
-    /// format string fails here.
-    /// Mutates: changing the separator from " / " to " - " fails this;
-    /// dropping leading-zero pad on seconds fails this.
-    func testMiniPlayer_formatTime_returnsExpectedMMSS() {
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(0), "0:00",
-                       "Zero seconds must format as 0:00 (preserves leading zero on seconds)")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(59), "0:59",
-                       "59 seconds must format as 0:59")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(60), "1:00",
-                       "60 seconds must format as 1:00 — minute rollover")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(125), "2:05",
-                       "125 seconds must format as 2:05")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(3600), "1:00:00",
-                       "1 hour must format as 1:00:00 — hour rollover adds the H field")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(3725), "1:02:05",
-                       "1h 2m 5s must format as 1:02:05 — leading-zero pad on M field when hours present")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(-1), "--:--",
-                       "Negative inputs must format as --:-- (guard against malformed positions)")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(.nan), "--:--",
-                       "NaN must format as --:-- (guard against /0 in playbackProgress upstream)")
-        XCTAssertEqual(AudiobookMiniPlayerView.formatTime(.infinity), "--:--",
-                       "Infinity must format as --:-- (defensive against toolkit edge cases)")
-    }
-}
